refactor(library): modularize Schema component and extract sub-components
