+++ library/src/components/Schema/AdditionalProperties.tsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import { SchemaInterface } from '@asyncapi/parser';
+
+import { Schema } from './Schema';
+import { SchemaHelpers } from '../../helpers';
+
+interface AdditionalPropertiesProps {
+    schema: SchemaInterface;
+}
+
+export const AdditionalProperties: React.FunctionComponent<
+    AdditionalPropertiesProps
+> = ({ schema }) => {
+    if (
+        schema.extensions().get(SchemaHelpers.extRenderAdditionalInfo)?.value() ===
+        false
+    ) {
+        return null;
+    }
+
+    const type = schema.type();
+    if (!type?.includes('object')) {
+        return null;
+    }
+
+    const additionalProperties = schema.additionalProperties();
+    if (additionalProperties === true || additionalProperties === undefined) {
+        return (
+            <p className="mt-2 text-xs text-gray-700">
+                Additional properties are allowed.
+            </p>
+        );
+    }
+    if (additionalProperties === false) {
+        return (
+            <p className="mt-2 text-xs text-gray-700">
+                Additional properties are <strong>NOT</strong> allowed.
+            </p>
+        );
+    }
+    return (
+        <Schema schemaName="Additional properties:" schema={additionalProperties} />
+    );
+};
