+++ library/src/components/Schema/SchemaContext.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export const SchemaContext = React.createContext({
+    reverse: false,
+    deepExpanded: false,
+});
